Adds proper checks for all installed files
This makes overwriting someones install files impossible. Instead it is
up to the user to manually remove the files themselves.

+++ install.sh
@@ -1,18 +1,31 @@
 #!/bin/bash
 
+# To be used at initial setup only. If the files already exist a simple warning
+# is issued and no action taken. It is up to the user to delete the file and
+# rerun this script to properly install
+
 # Get current dir (so run this script from anywhere)
 export DOTFILES_DIR="$( cd "$( dirname "${BASH_SOURCE[0]}" )" && pwd )"
 
 # Update dotfiles itself first
 [ -d "$DOTFILES_DIR/.git" ] && git --work-tree="$DOTFILES_DIR" --git-dir="$DOTFILES_DIR/.git" pull origin master
 
-# symlinks for standard files
-ln -sfv "$DOTFILES_DIR/.bashrc" ~
-ln -sfv "$DOTFILES_DIR/.gitconfig" ~
+# Setup symlinks - if they already exist just print a warning
+if [ -f ~/.bashrc ]; then
+    echo "WARNING: ~/.bashrc already exists: no action taken"
+else
+    ln -sfv "$DOTFILES_DIR/.bashrc" ~
+fi
+
+if [ -f ~/.gitconfig ]; then
+    echo "WARNING: ~/.gitconfig already exists: no action taken"
+else
+    ln -sfv "$DOTFILES_DIR/.gitconfig" ~
+fi
 
 # vim setup
 if [ -d ~/.vim ]; then
-  echo "~/.vim already exist - ignoring vim setup"
+  echo "WARNING: ~/.vim already exist: no action taken"
 else
   ln -sfv "$DOTFILES_DIR/.vim" ~
